Suppress "AUTOSAR C++14 M5-0-3" rule finding
This rule states: "A cvalue expression shall not be implicitly converted to a different underlying type.".

GIT_ORIGIN_SPP_REV_ID: 175dcf89534ec2ea1433e76e50a96ff08304eea8

+++ score/message_passing/qnx_dispatch/qnx_dispatch_engine.cpp
@@ -209,7 +209,7 @@ void QnxDispatchEngine::RegisterPosixEndpoint(PosixEndpointEntry& endpoint) noex
 
     if (posix_receive_buffer_.size() < endpoint.max_receive_size)
     {
-        posix_receive_buffer_.resize(endpoint.max_receive_size);
+        posix_receive_buffer_.resize(static_cast<std::size_t>(endpoint.max_receive_size));
     }
 
     score::cpp::ignore = os_resources_.dispatch->select_attach(
